Add files via upload
This will allow you to change attributes of anything on your web page. Extremely useful to change images to avoid pages reloads.
This inlcudes an HTML and JS part.
NOTE: In case of adding multiple items to the switcher I highly recommend building promises in order to have a cleaner and stable code.